+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 2017-03-16
+* `[SEMVER_MAJOR]` Removed `uriTemplate` param - document URIs will now always be determined by the provided `getUri` function
 * `[SEMVER_PATCH]` Fixed a bug that caused properties with a boolean `false` spec to still be included in transformed output 
 
 ## 2017-03-12